pyproject.toml
[project]
name = "neurobids-flow"
version = "1.0.0"
description = "A modular graphical framework for standardizing multi-source EEG recordings to BIDS-EEG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "mne>=1.11.0",
    "mne-bids>=0.18.0",
    "pandas>=3.0.1",
    "pyxdf>=1.17.3",
    "pyyaml>=6.0.3",
]
[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
[tool.setuptools.package-dir]
"" = "src"
[project.scripts]
neurobids-flow = "neurobids_flow.cli:cli"